Перенести часть компонентов и логики из LayersTable в LayersPage
